// src/app/module/auth.interface.ts
import type { Role } from "../../generated/prisma/enums";

export interface IRegisterUserPayload {
	name: string;
	email: string;
	password: string;
	phone?: string;
	role?: "CUSTOMER" | "TECHNICIAN";
}

export interface IVerifyEmailPayload {
	email: string;
	otp: string;
}

export interface ILoginUserPayload {
	email: string;
	password: string;
}

export interface IGoogleLoginPayload {
	idToken: string;
}

export interface IForgotPasswordPayload {
	email: string;
}

export interface IResetPasswordPayload {
	email: string;
	otp: string;
	newPassword: string;
}

export interface IRequestUser {
	userId: string;
	name: string;
	email: string;
	role: Role;
}